Huge Update Fix a few things.

+++ FullstackProject/Model/Membership.cs
@@ -8,14 +8,12 @@ namespace FullstackProject.Model
         public int MembershipId { get; set; }
         public int? UserId { get; set; }
         public string? Type { get; set; }
-        public float? DiscountRate { get; set; }
+        public double? DiscountRate { get; set; }
         public int? Price { get; set; }
         public string? Description { get; set; }
 
         // Navigation property for User
         public virtual User? User { get; set; }
-
-        
-            }
-        }
+    }
+}
